Use ENV variables to set database connection
To be able to Dockerize for the dxw hosting platform, the database connection needs to be set via ENV variables.

I have chosen not to set this via a databse URL parameter, as that makes it clumsier to change just parts of the defaults, such as host name, without also overwriting the database name between environments.

I have updated the docker-compose-env.sample to reflect this. With the provided defaults, it will work locally either running it on a default local development setup, or using docker-compose. For production, DB_HOST_NAME, DB_USERNAME and DB_PASSWORD will probably all have to be set appropiately.

Because this fixes the databse connection, I can uncomment the docker-entrypoint.sh docker setup, which automatically sets up the database when you run `docker-compose up`