Validate presence of application number on development
It's not entirely clear yet which of these fields are mandatory to be able to create a development. Do developments start getting logged/discussed before the have an application number/site address/proposal set? Probably not, but we want to keep it flexible until we find out.

However, if you create a development without an application number, then things start breaking, so let's require that.