Turn off auditing on seeds
When running the seeds, the system requires would create an audit log for any dwellings added to a dwelling that isn't in a draft state. This would also error because we're not adding an audit_comment to these changes.

I've turned off auditing for these seeds.
I also added seed.rb to the rubocop ignore list, as it was complaining about the block being too large, which is OK in this case.