Add address and registered provider fields to dwellings
These are the fields that developers will fill in once a development is completed.

We don't want to generate an audit log when a developer does this, so exclude those from the auditing, which then also requires us to do the same monkey patch as we did on Development to stop it generating a validation error when they are changed without an audit comment.