Allow Southwark user to request a developer completion response
If a development is 'completed', and contains any dwellings that are within s106 that don't have an address or registered provider assigned, then we ask the southwark user to request this from the developer, with the URL provided.

If the development is `completed`, and this has all been filled in, then let them know that it has.