Use fork of Audited gem
We had two bugs with the audited gem, and I've opened PRs for both of them:

https://github.com/collectiveidea/audited/pull/522
https://github.com/collectiveidea/audited/pull/523

In the meantime, while we're waiting for them to hopefully be approved and merged, I have created my own fork with a branch that has both of those changes in.

This means that we no longer need the monkey pathes in the dwelling and development models. Also update the specs to reflect the new, fixed error messages.