Add concept of Development's primary application number
A development can now support having multiple application numbers through muliple planning applications.

However, this isn't the common case. Usually they will only have one. And if they have more than one, there'll still be an idea of a "primary" one. For now, we'll just use the first planning application that was created, but this leaves room in the future for us to change which one we use.

Then update all views to use this instead.