Associate planning application with audit change
We want our users to be able to link a change to an agreed dwelling to the planning application that the change was agreed in.

This requires us to add the association to the audit model, so we're creating our own affordable housing audit class.

Audited doesn't however let us assign the value of the association in any official way, so we're back to monkey patching the `write_audit` method to get it to work.

Now we have that working, we add the ability to choose the planning application when creating or editing dwellings on an agreed development.