Stop postgres error causing all builds to fail
The PostgreSQL Docker Team pushed a breaking change to stop users
running their production databases without a password. This caused the
test build to fail with an error.

Adding the `POSTGRES_HOST_AUTH_METHOD: trust` environment variable to
the postgres container should fix this. A password is not required in
this case as this database will only ever contain test data.