Add support for dwellings that do not have a development
Changes to the schema and validation to allow a dwelling to not belong to a development.

Various parts of the application would break were such a dwelling exposed to it, so I have added a default_scope to hide them by default. This is a temporary fix, and should be removed in the future.

We do not want to audit changes made to a dwelling that does not belong to a dwelling, as I'm not sure what that would do.

Habitable Rooms, Bedrooms and Developer Reference ID are all no longer mandatory if dwelling si not part of a development.